# [PERF] PromQL: Speed up PromQL to string conversion