Goes faster due to reduced memory allocation, and skipping parsing of the `%s` format string.
Replace `Sprintf` with either string concat or writing to a `bytes.Buffer`.

This is not a frequent operation in Prometheus itself, but ends up getting called a lot in downstream project Mimir.

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-notes block below.
Otherwise, please describe what should be mentioned in the CHANGELOG. Use the following prefixes:
[FEATURE] [ENHANCEMENT] [PERF] [BUGFIX] [SECURITY] [CHANGE]
Refer to the existing CHANGELOG for inspiration:  https://github.com/prometheus/prometheus/blob/main/CHANGELOG.md
If you need help formulating your entries, consult the reviewer(s).
-->
```release-notes
NONE
```

## Benchmarks

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql/parser
cpu: Apple M2
                                                            │ e44ee2f18     │           153cdb2b0                │
                                                            │    sec/op     │   sec/op     vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8       425.6n ±  7%   163.7n ± 3%  -61.54% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8      1846.0n ±  2%   914.9n ± 1%  -50.44% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8      1653.5n ±  3%   935.7n ± 9%  -43.41% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8       3.502µ ± 10%   1.973µ ± 2%  -43.65% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8    3.465µ ±  7%   2.027µ ± 4%  -41.49% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8       976.8n ± 19%   488.9n ± 3%  -49.95% (p=0.002 n=6)
geomean                                                        1.577µ         806.0n       -48.90%

                                                            │ e44ee2f18     │            153cdb2b0                │
                                                            │     B/op      │     B/op      vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8         240.0 ± 0%     104.0 ± 0%  -56.67% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8        1457.0 ± 0%     944.0 ± 0%  -35.21% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8         856.0 ± 0%     645.0 ± 0%  -24.65% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8       3.434Ki ± 0%   2.492Ki ± 0%  -27.42% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8    3.386Ki ± 0%   2.445Ki ± 0%  -27.78% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8         720.0 ± 0%     344.0 ± 0%  -52.22% (p=0.002 n=6)
geomean                                                        1.147Ki          719.9       -38.71%

                                                            │ e44ee2f18     │           153cdb2b0               │
                                                            │   allocs/op   │ allocs/op   vs base               │
ExprString/sum_by(code)_(task:errors:rate10s{job="s...-8        12.000 ± 0%   4.000 ± 0%  -66.67% (p=0.002 n=6)
ExprString/max(_100_*_(1_-_avg_by(instance)_(irate(...-8         34.00 ± 0%   18.00 ± 0%  -47.06% (p=0.002 n=6)
ExprString/http_requests_total{job="api-server",_gr...-8         30.00 ± 0%   19.00 ± 0%  -36.67% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...-8         58.00 ± 0%   32.00 ± 0%  -44.83% (p=0.002 n=6)
ExprString/sum_by_(pod)_((kube_pod_container_status...#01-8      58.00 ± 0%   32.00 ± 0%  -44.83% (p=0.002 n=6)
ExprString/label_replace(testmetric,_"dst",_"destin...-8         22.00 ± 0%   13.00 ± 0%  -40.91% (p=0.002 n=6)
geomean                                                          31.11        16.22       -47.86%
```